--- a/parser/checkpoint.go
+++ b/parser/checkpoint.go
@@ -39,14 +39,65 @@ func (c *CheckpointEvent) Parse(r reader.Reader, classes ClassMap, cpools PoolMa
 		if err != nil {
 			return fmt.Errorf("unable to parse constant pool class: %w", err)
 		}
-		cm, ok := cpools[int(classID)]
-		if !ok {
-			cpools[int(classID)] = &CPool{Pool: make(map[int]ParseResolvable)}
-			cm = cpools[int(classID)]
-		}
 		m, err := r.VarInt()
 		if err != nil {
 			return fmt.Errorf("unable to parse constant pool's number of constants: %w", err)
+		}
+		cm, ok := cpools[int(classID)]
+		if !ok {
+			cpools[int(classID)] = &CPool{Pool: make(map[int]ParseResolvable, m)}
+			cm = cpools[int(classID)]
+		}
+		class, ok := classes[int(classID)]
+		if !ok {
+			return fmt.Errorf("unexpected class %d", classID)
+		}
+		var (
+			results            []ParseResolvable
+			preAllocateResults = true
+			contantsSlice      = getConstantsSlice(int(m), class.numConstants)
+		)
+		// preallocate common class in async-profiler
+		results = make([]ParseResolvable, m)
+		switch class.Name {
+		case "java.lang.Thread":
+			threads := make([]Thread, m)
+			for i := range threads {
+				threads[i].constants = contantsSlice[i]
+				results[i] = &threads[i]
+			}
+		case "jdk.types.StackTrace":
+			stackTraces := make([]StackTrace, m)
+			for i := range stackTraces {
+				stackTraces[i].constants = contantsSlice[i]
+				results[i] = &stackTraces[i]
+			}
+		case "jdk.types.Method":
+			methods := make([]Method, m)
+			for i := range methods {
+				methods[i].constants = contantsSlice[i]
+				results[i] = &methods[i]
+			}
+		case "java.lang.Class":
+			classes := make([]Class, m)
+			for i := range classes {
+				classes[i].constants = contantsSlice[i]
+				results[i] = &classes[i]
+			}
+		case "jdk.types.Package":
+			packages := make([]Package, m)
+			for i := range packages {
+				packages[i].constants = contantsSlice[i]
+				results[i] = &packages[i]
+			}
+		case "jdk.types.Symbol":
+			symbols := make([]Symbol, m)
+			for i := range symbols {
+				symbols[i].constants = contantsSlice[i]
+				results[i] = &symbols[i]
+			}
+		default:
+			preAllocateResults = false
 		}
 		// TODO: assert m is small enough
 		for j := 0; j < int(m); j++ {
@@ -54,12 +105,28 @@ func (c *CheckpointEvent) Parse(r reader.Reader, classes ClassMap, cpools PoolMa
 			if err != nil {
 				return fmt.Errorf("unable to parse contant's index: %w", err)
 			}
-			v, err := ParseClass(r, classes, cpools, classID)
-			if err != nil {
-				return fmt.Errorf("unable to parse constant type %d: %w", classID, err)
+
+			var v ParseResolvable
+			if preAllocateResults {
+				v = results[j]
+				err = v.Parse(r, classes, cpools, class)
+			} else {
+				v, err = ParseClass(r, classes, cpools, classID)
+				if err != nil {
+					return fmt.Errorf("unable to parse constant type %d: %w", classID, err)
+				}
 			}
 			cm.Pool[int(idx)] = v
 		}
 	}
 	return nil
+}
+
+func getConstantsSlice(size int, num int) [][]constant {
+	constants := make([]constant, size*num)
+	contantsSlice := make([][]constant, size)
+	for i := range contantsSlice {
+		contantsSlice[i] = constants[i*num : (i+1)*num][:0]
+	}
+	return contantsSlice
 }