--- a/parser/checkpoint.go
+++ b/parser/checkpoint.go
@@ -101,6 +101,10 @@ func (c *CheckpointEvent) Parse(r reader.Reader, classes ClassMap, cpools PoolMa
 			var pointerToStackFrames []*StackFrame
 			indexPointerToStackFrames := 0
 			getPointerToStackFrames := func(n int) []*StackFrame {
+				if n > int(m) {
+					pointerToStackFrames = make([]*StackFrame, n)
+					return pointerToStackFrames
+				}
 				if indexPointerToStackFrames+n > len(pointerToStackFrames) {
 					pointerToStackFrames = make([]*StackFrame, m)
 					indexPointerToStackFrames = 0