--- a/parser/metadata.go
+++ b/parser/metadata.go
@@ -48,11 +48,13 @@ func (s *SettingMetadata) SetAttribute(key, value string) error {
 func (s SettingMetadata) AppendChild(string) Element { return nil }
 
 type FieldMetadata struct {
-	Class        int64
-	Name         string
-	ConstantPool bool
-	Dimension    int32
-	Annotations  []AnnotationMetadata
+	Class                int64
+	Name                 string
+	ConstantPool         bool
+	Dimension            int32
+	Annotations          []AnnotationMetadata
+	isBaseType           bool
+	parseBaseTypeAndDrop func(r reader.Reader) error
 }
 
 func (f *FieldMetadata) SetAttribute(key, value string) (err error) {