--- a/parser/parser.go
+++ b/parser/parser.go
@@ -6,14 +6,14 @@ import (
 )
 
 func Parse(r io.Reader) ([]Chunk, error) {
-	return ParseWithOptions(r, &ChunkParseOptions{})
+	return ParseWithOptions(r, &ChunkParseOptions{}, true)
 }
 
-func ParseWithOptions(r io.Reader, options *ChunkParseOptions) ([]Chunk, error) {
+func ParseWithOptions(r io.Reader, options *ChunkParseOptions, unsafeByteToString bool) ([]Chunk, error) {
 	var chunks []Chunk
 	for {
 		var chunk Chunk
-		err := chunk.Parse(r, options)
+		err := chunk.Parse(r, options, unsafeByteToString)
 		if err == io.EOF {
 			return chunks, nil
 		}