reduce 99% memory allocation and save 55% cpu.

go version:
```
go version go1.18.9 linux/amd64
```

benchstat:
```
goos: linux
goarch: amd64
pkg: github.com/pyroscope-io/jfr-parser/parser
cpu: Intel(R) Xeon(R) Platinum 8275CL CPU @ 3.00GHz
                       │   old.txt    │               new.txt               │
                       │    sec/op    │   sec/op     vs base                │
Parse/example-8          20.914m ± 1%   8.647m ± 1%  -58.65% (p=0.000 n=10)
Parse/async-profiler-8   118.61m ± 1%   40.10m ± 0%  -66.19% (p=0.000 n=10)
geomean                   49.81m        18.62m       -62.61%

                       │    old.txt    │               new.txt                │
                       │     B/op      │     B/op      vs base                │
Parse/example-8           6.424Mi ± 0%   5.276Mi ± 0%  -17.87% (p=0.000 n=10)
Parse/async-profiler-8   42.484Mi ± 0%   4.380Mi ± 0%  -89.69% (p=0.000 n=10)
geomean                   16.52Mi        4.807Mi       -70.90%

                       │    old.txt    │               new.txt               │
                       │   allocs/op   │  allocs/op   vs base                │
Parse/example-8          169.264k ± 0%   1.193k ± 0%  -99.30% (p=0.000 n=10)
Parse/async-profiler-8   439964.0 ± 0%    938.0 ± 0%  -99.79% (p=0.000 n=10)
geomean                    272.9k        1.058k       -99.61%
```