@zdyj3170101136 2023-08-02
> That's quite a lot of changes. Could you describe what the changes are?

1, the reader use binary.Read to get int and string, let the memory escape to heap.
 commit https://github.com/grafana/jfr-parser/pull/22/commits/39df228213cf2f7c3e6f9cfd1c47ff51e12ef87d  fix this.
2, in types.go, even for base type like int, it will malloc a Int type, which pointer to int. commit https://github.com/grafana/jfr-parser/pull/22/commits/e587646bda0001d7a48dba80ce8aad2b89ec2ea0 fix it, add argument reader.Reader to function parseField, change toLong(p Parseable) to toLong(r reader.Reader).
3, batch allocate stackframe and constants, commit https://github.com/grafana/jfr-parser/pull/22/commits/2aa0b3327c30b2ede2013f7a2ff810b4c5561166 and https://github.com/grafana/jfr-parser/pull/22/commits/1dcd969e77f74b0583a672ee5c99337ffa5a84fa fix it.
4, in building cpool, perallocate each class type. commit https://github.com/grafana/jfr-parser/pull/22/commits/5660ef6109c5bc43f8a7d9289b2db62a245066d8 and https://github.com/grafana/jfr-parser/pull/22/commits/a0d5347e6a4c39ed198e14b0208609f293a5953b fix it.
5, perallocate stackframe, it is special because cpool could not malloc it directly from ParseClass but from parsefields. commit https://github.com/grafana/jfr-parser/pull/22/commits/a1b27a85a272294a4a4770a574ecd38c23896b51 and commit https://github.com/grafana/jfr-parser/pull/22/commits/550121da815e9b861cabc29aec75bca1b5097d58 fix it.
6, let chunk implement chunk.Next() interface(), so we could cache common event_type, commit https://github.com/grafana/jfr-parser/pull/22/commits/a526cc1b3e6ecaa1caa73c13babf480c13b0953b fix it.
7, the ClassMap store ClassMetaData as value in default, this lead to a lot of duffcopy, now the ClassMap store pointer to ClassMetaData. commit https://github.com/grafana/jfr-parser/pull/22/commits/7151b79c6a490eedef4da7d2427b60e5eba75a27 fix it.
8, parseFields use for _, v range over class.fields, lots of copy. fix this in https://github.com/grafana/jfr-parser/pull/22/commits/91d9ca7447f1c387edef4c62474035f9468ecafb.
9, there a lot of type which only have base type like string, int. but also implement Resolve interface{}, change it to return directly. commit https://github.com/grafana/jfr-parser/pull/22/commits/d66951dc9605fde3726821e643c134ab351cb7cb fix this.