@korniltsev 2023-08-02
Thanks a lot for an explanation and for the PR
I will look into it soon.

Could you  please also add benchmark results to the PR description, if possible using benchstat