@zdyj3170101136 2023-08-02
> reduce 99% memory allocation and save 55% cpu.
> 
> benchstat:
> 
> ```
> goos: linux
> goarch: amd64
> pkg: github.com/pyroscope-io/jfr-parser/parser
> cpu: Intel Xeon Processor (Icelake)
>          │   old.txt   │               new.txt               │
>          │   sec/op    │   sec/op     vs base                │
> Parse-16   26.00m ± 9%   11.52m ± 3%  -55.67% (p=0.000 n=10)
> 
>          │   old.txt    │               new.txt                │
>          │     B/op     │     B/op      vs base                │
> Parse-16   6.426Mi ± 0%   5.278Mi ± 0%  -17.86% (p=0.000 n=10)
> 
>          │    old.txt    │               new.txt               │
>          │   allocs/op   │  allocs/op   vs base                │
> Parse-16   169.265k ± 0%   1.193k ± 0%  -99.30% (p=0.000 n=10)
> ```



> Thanks a lot for an explanation and for the PR I will look into it soon.
> 
> Could you please also add benchmark results to the PR description, if possible using benchstat

already add benchmark.