@zdyj3170101136 2023-08-02
more detailed benchstat.
each number repersent operation as https://github.com/grafana/jfr-parser/pull/22#issuecomment-1661437669 said.
each compared to last operation.
seems 8 and 9 not really have enhancement.
operation 6 have small enhancement cause the example jfr file is small, only contains small number of event.
1, 
```
goos: linux
goarch: amd64
pkg: github.com/pyroscope-io/jfr-parser/parser
cpu: Intel Xeon Processor (Icelake)
         │   old.txt   │               old.txt1               │
         │   sec/op    │    sec/op     vs base                │
Parse-16   26.00m ± 9%   21.86m ± 14%  -15.91% (p=0.003 n=10)

         │   old.txt    │              old.txt1               │
         │     B/op     │     B/op      vs base               │
Parse-16   6.426Mi ± 0%   6.139Mi ± 0%  -4.46% (p=0.000 n=10)

         │   old.txt   │              old.txt1              │
         │  allocs/op  │  allocs/op   vs base               │
Parse-16   169.3k ± 0%   154.1k ± 0%  -8.94% (p=0.000 n=10)
```
2, 
```
goos: linux
goarch: amd64
pkg: github.com/pyroscope-io/jfr-parser/parser
cpu: Intel Xeon Processor (Icelake)
         │   old.txt1   │               old.txt2               │
         │    sec/op    │    sec/op     vs base                │
Parse-16   21.86m ± 14%   17.51m ± 10%  -19.89% (p=0.000 n=10)

         │   old.txt1   │              old.txt2               │
         │     B/op     │     B/op      vs base               │
Parse-16   6.139Mi ± 0%   5.872Mi ± 0%  -4.35% (p=0.000 n=10)

         │   old.txt1   │              old.txt2               │
         │  allocs/op   │  allocs/op   vs base                │
Parse-16   154.14k ± 0%   93.08k ± 0%  -39.61% (p=0.000 n=10)
```
3, 
```
goos: linux
goarch: amd64
pkg: github.com/pyroscope-io/jfr-parser/parser
cpu: Intel Xeon Processor (Icelake)
         │   old.txt2   │              old.txt3              │
         │    sec/op    │   sec/op     vs base               │
Parse-16   17.51m ± 10%   16.05m ± 9%  -8.33% (p=0.029 n=10)

         │   old.txt2   │               old.txt3               │
         │     B/op     │     B/op      vs base                │
Parse-16   5.872Mi ± 0%   4.537Mi ± 0%  -22.73% (p=0.000 n=10)

         │  old.txt2   │              old.txt3               │
         │  allocs/op  │  allocs/op   vs base                │
Parse-16   93.08k ± 0%   60.35k ± 0%  -35.17% (p=0.000 n=10)
```
4, 
```
goos: linux
goarch: amd64
pkg: github.com/pyroscope-io/jfr-parser/parser
cpu: Intel Xeon Processor (Icelake)
         │  old.txt3   │            old.txt4            │
         │   sec/op    │    sec/op     vs base          │
Parse-16   16.05m ± 9%   15.18m ± 10%  ~ (p=0.190 n=10)

         │   old.txt3   │              old.txt4               │
         │     B/op     │     B/op      vs base               │
Parse-16   4.537Mi ± 0%   4.525Mi ± 0%  -0.27% (p=0.000 n=10)

         │  old.txt3   │              old.txt4              │
         │  allocs/op  │  allocs/op   vs base               │
Parse-16   60.35k ± 0%   56.42k ± 0%  -6.51% (p=0.000 n=10)
```
5, 
```
goos: linux
goarch: amd64
pkg: github.com/pyroscope-io/jfr-parser/parser
cpu: Intel Xeon Processor (Icelake)
         │   old.txt4   │              old.txt5               │
         │    sec/op    │   sec/op     vs base                │
Parse-16   15.18m ± 10%   11.01m ± 4%  -27.46% (p=0.000 n=10)

         │   old.txt4   │               old.txt5               │
         │     B/op     │     B/op      vs base                │
Parse-16   4.525Mi ± 0%   5.487Mi ± 0%  +21.27% (p=0.000 n=10)

         │   old.txt4   │              old.txt5               │
         │  allocs/op   │  allocs/op   vs base                │
Parse-16   56.418k ± 0%   2.684k ± 0%  -95.24% (p=0.000 n=10)
```
6
```
goos: linux
goarch: amd64
pkg: github.com/pyroscope-io/jfr-parser/parser
cpu: Intel Xeon Processor (Icelake)
         │  old.txt5   │              old.txt6              │
         │   sec/op    │   sec/op     vs base               │
Parse-16   11.01m ± 4%   10.28m ± 6%  -6.64% (p=0.001 n=10)

         │   old.txt5   │              old.txt6               │
         │     B/op     │     B/op      vs base               │
Parse-16   5.487Mi ± 0%   5.282Mi ± 0%  -3.73% (p=0.000 n=10)

         │  old.txt5   │              old.txt6               │
         │  allocs/op  │  allocs/op   vs base                │
Parse-16   2.684k ± 0%   1.236k ± 0%  -53.95% (p=0.000 n=10)
```
7
```
goos: linux
goarch: amd64
pkg: github.com/pyroscope-io/jfr-parser/parser
cpu: Intel Xeon Processor (Icelake)
         │  old.txt6   │           old.txt7            │
         │   sec/op    │   sec/op     vs base          │
Parse-16   10.28m ± 6%   10.23m ± 4%  ~ (p=0.190 n=10)

         │   old.txt6   │              old.txt7               │
         │     B/op     │     B/op      vs base               │
Parse-16   5.282Mi ± 0%   5.275Mi ± 0%  -0.14% (p=0.000 n=10)

         │  old.txt6   │              old.txt7              │
         │  allocs/op  │  allocs/op   vs base               │
Parse-16   1.236k ± 0%   1.193k ± 0%  -3.48% (p=0.000 n=10)
```
8, 9 
```
goos: linux
goarch: amd64
pkg: github.com/pyroscope-io/jfr-parser/parser
cpu: Intel Xeon Processor (Icelake)
         │   old.txt7   │           old.txt8            │
         │    sec/op    │   sec/op     vs base          │
Parse-16   10.225m ± 4%   9.987m ± 4%  ~ (p=0.280 n=10)

         │   old.txt7   │            old.txt8            │
         │     B/op     │     B/op      vs base          │
Parse-16   5.275Mi ± 0%   5.275Mi ± 0%  ~ (p=0.579 n=10)

         │  old.txt7   │           old.txt8            │
         │  allocs/op  │  allocs/op   vs base          │
Parse-16   1.193k ± 0%   1.193k ± 0%  ~ (p=0.628 n=10)
```